perf(muf): FP32 SH force backprop + combined B-spline — 7.2% GPU speedup
V7 Kernel: FP32 accumulate_f12_one_f32 for SH force backprop.
- FP32 inner loop leverages 128 FP32 units/SM vs 2-4 FP64 units on RTX 5090
- Combined muf_eval_bspline_smooth_both saves one cos()+sin() per neighbor
- 2B kernel: combined B-spline + register force accumulation
- All-L W_eff precompute in shared memory (single barrier)

Performance (66,990 SiGe, RTX 5090, pure GPU):
  Fused kernel: 2.706ms → 2.465ms (-8.9%)
  Total GPU:    3.623ms → 3.349ms (-7.6%)
  Speed:        17.40M → 18.66M atoms·steps/s (+7.2%)

Energy verified: -263907.70344 eV (identical to FP64 baseline)

+++ src/force/muf.cuh
@@ -58,7 +58,7 @@ struct MUF_Data {
   GPU_Vector<int> NN;           // neighbor count per atom
   GPU_Vector<int> NL;           // flat neighbor list (GPU)
 
-  GPU_Vector<double> moments;   // type-channel descriptors
+  GPU_Vector<float> moments;    // type-channel descriptors (FP32: halved memory)
   // [N * num_types * K * num_sh_terms]
   GPU_Vector<double> dE_dA;     // gradient adjoint (same size as moments)
   GPU_Vector<double> energy_3b; // per-atom 3B energy
